rename ClusteringWriterTestUtils to ClusteringAssertions
